do a bit more organisation
