+++ util/now.py
@@ -1,0 +1,6 @@
+import pytz
+import datetime
+
+def now():
+    tz = pytz.timezone('AUSTRALIA/Adelaide')
+    return datetime.now(tz).strftime("-%H.%M.%S %d.%m.%y")
